Added Data Annotations to DTO objects related to the CreateParkingSpace view.
